fix(misc): add namespace, fixed class references
